requirements.txt
appdirs==1.4.3
asn1crypto==0.23.0
awscli==1.11.50
Babel==2.3.4
BeautifulSoup==3.2.1
beautifulsoup4==4.5.3
bootstrap-admin==0.3.7.1
boto==2.45.0
botocore==1.5.13
braintree==3.39.0
bson==0.5.0
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
click==6.7
colorama==0.3.7
colorlog==2.10.0
cryptography==2.0.3
dj-database-url==0.4.2
Django==1.8.4
django-annoying==0.10.3
django-bower==5.2.0
django-cities==0.5.0.5
django-cors-headers==2.0.2
django-countries==4.5
django-crispy-forms==1.6.1
django-filter==0.15.0
django-generic-positions==0.2.2
django-hvad==1.8.0
django-libs==1.67.12
django-likert-field==0.2.0
django-model-utils==3.0.0
django-phonenumber-field==1.1.0
django-registration-redux==1.2
django-review==1.9.6
django-scheduler==0.8.4
django-storages-redux==1.3.2
django-user-media==1.2.4
django-voting==0.2
djangorestframework==3.6.2
docutils==0.13.1
easy-thumbnails==2.4.1
enum34==1.1.6
Flask==0.12
funcsigs==1.0.2
future==0.16.0
futures==3.0.5
fuzzywuzzy==0.15.0
GDAL==2.2.1
geoip2==2.5.0
geos==0.2.1
get==0.0.0
googlemaps==2.5.1
gunicorn==19.6.0
icalendar==3.11.4
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.4
jmespath==0.9.1
lxml==3.7.3
MarkupSafe==0.23
maxminddb==1.3.0
mock==2.0.0
ndg-httpsclient==0.4.3
packaging==16.8
pbr==1.10.0
Pillow==2.9.0
post==0.0.0
psycopg2==2.6.2
public==0.0.0
pyasn1==0.2.2
pycparser==2.18
pyjade==4.0.0
pyOpenSSL==17.3.0
pyparsing==2.1.10
python-dateutil==2.6.0
python-http-client==2.2.1
pytz==2016.10
PyYAML==3.12
query-string==0.0.0
request==0.0.0
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.10
sendgrid==3.6.3
setupfiles==0.0.0
simplejson==3.10.0
six==1.10.0
South==1.0.2
srvup-comments==1.0.6
sure==1.4.1
swapper==1.0.0
tqdm==4.11.2
udemy-dl==0.2.1
unicode-slugify==0.1.3
Unidecode==0.4.19
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.11.15
whitenoise==3.3.0
youtube-dl==2016.12.12
